Starting conversion to correct monadic style.
The previous solution for the arbitrary-arity CPSer used a style that
was very, *very* close to a monadic one, but wasn't quite
right. Rather, it conflated the serious expression transformer, S,
with the "bind" operator for the continuation monad. The changes here
make the continuation monad explicit, using returnK and bindK. This
also changes S so that it returns a proper continuation
monad. Additionally, S only takes an expression and no longer requires
the current continuation (this is due to the inclusion of bindK which
does the work of constructing the continuation).

Early work has also been done to clean up the bind forms using letMK
and letMK* to simulate Haskell's "do" notation, but it seems there are
some kinks. With the additional of letMK and letMK* into the code, it
no longer works as expected.