Refactoring tramp to work accurately.
Previously, tramp used a pseudo-CPS method of re-writing the code for
trampolining to set the "done" variable. Due to poor design and
reasoning, this led to what could not be called a
"correctness-preserving transformation." Rather, it could (and should)
have been called an "incorrectness-introducing transformation."

Now tramp examines function applications to determine whether the
operator-position is the continuation argument for the current
function. Such function applications are known to be value-producing
contexts, and are thus the locations where the "done" variable should
be set.