Removing "done" and "kv" parameters from tramp and helper functions.
Previously tramp passed around the "done" variable and "kv"
continuation variable. This was done to allow the tramp-app helper to
detect when the continuation was being applied in the body of a
computation and choose that expression to transform into setting the
"done" variable (also hence passing the "done" variable).

Testing showed that this was not producing correct output and passing
the parameters became unnecessary. They are being removed, which also
removes the need for the "tramp-main" helper since the additional
parameters are unnecessary.