Changing the method of setting the "done" variable.
Previously, tramp transformed function application expressions in
which the continuation was being applied to additionally set the
"done" variable, indicating that the trampolined computation was
complete. Testing the Ackermann function showed that this didn't work
correctly when the continuation contained a call to a thunkified
function. That is, setting the "done" variable indicated the
computation was complete, but it actually required more bounces on the
trampoline (that's really hard to type seriously with a straight
face).

This has been changed such that the initial continuation to a function
call is composed such that it flips the "done" variable on
invocation. Testing shows that this corrects the issues with the
Ackermann function.