Converting abstract-k to use sets of continuation variables.
Previously abstract-k abstracted over one continuation argument at a
time. This became problematic when dealing with continuation
application for continuations other than the one currently being
abstracted. The problem stemmed from the fact that all
non-continuation function applications necessarily have a continuation
as their last argument, which must be handled specially for
abstraction.

This broke because continuation application for continuations other
than the current one were treated as normal function application
(which they're not). Thus, the continuation argument (a value) would
be treated as a continuation, passed to the handler for continuation
arguments and blow up.

This required that the abstraction algorithm consider all continuation
variables at once. It now carries a set of the arguments and abstracts
over the application of each one as they're encountered, which solves
the problem for simple tests.