Implementing all the language forms in terms of the new CPS protocols.
Previously the system used a single CPS protocol (aptly named PCps)
that tried to encapsulate CPS for all expressions. This has been
teased out into two protocols, PCpsSrs and PCpsTriv (see commit
e01f4c7). These separately define the CPS transformation for
expressions considered "serious" or "trivial" with respect to the
Olivier-style CPS algorithm.

A major advantage of this approach is that it maps the logic for the
algorithm onto the expressions much more closely than
before. Additionally, it removes the need for the 'triv?' predicate
that PCps included, since being trivial or serious is encoded in which
protocol the expression implements.

This has a slightly unfortunate side-effect of creating some more code
duplication. Specific examples include handling the body of 'fn'
expressions and the consequent and alternative of 'if'
expressions. These correspond to calls to 'expr' in the original,
pattern-matching version of the TCO compiler (known as 'E' in Olivier
Danvy's "A First-Order One-Pass CPS Transformation").