Implementing 'let' as full language form for CTCO.
Modifying the existing implementation for 'let' expressions in CTCO to
handle it as user input to CTCO. This include removing the simple Let
record type used previously for the mini-passes run at the end of CTCO
and replacing it with LetCps, LetSrs, and LetTriv records according to
CPSed, "serious," and "trivial" 'let' expressions, respectively.

This includes implementing the PAbstractK, PThunkify, PWalkable
protocols for LetCps, PCpsSrs protocol for LetSrs, PCpsTriv protocol
for LetTriv, and PEmit protocol for all three.

Finally, extensive file-level documentation is being added for all
three record types.