Reducing thunkification of CTCO expressions and cleaning up docs.
Previously CTCO thunkified all function applications, including those
for continuation application. This supported greater recursion depth
for expressions which were not in fact tail-recursive (i.e. the
intuitive definition of factorial), but reduced the overall
performance of transformed expressions. This modification thus adds
better support for tail-recursive expressions at the expense of
non-tail-recursive ones (which were not to be supported in any case).