Changing representation of fn to match Clojure's version.
Previously, the CTCO representation of fn was a stripped down version
of the one used in Clojure. The underlying record can now represent
all the forms usable in the Clojure fn form.

Note that the full set of forms for fn is not yet supported by
CTCO. The underlying representation can accept it, but the parser
needs to be updated to handle it correctly.