Adding a nifty extend-group macro
Many of the CTCO expression records have common implementations for
their protocols and make use of extend...repeatedly. Tonight I finally
recognized that this was screaming for a macro, and it turned out that
macro didn't exist. Well now it does, in the form of extend-group in
util.clj. It's not actually used yet, but should be propogated through
the expression record definitions.