Removing the Defn record type from CTCO.
Previously the compiler represented (almost) each supported language
form as a separate record type. Steps are being taken to strip back
the number of internal record types to use few more than what are
considered the "core forms" of the Clojure language. In order to
support a wide array of expressions, it falls to the parser to
reexpress those expressions in terms of semantically equivalent core
forms.

The removal of the Defn record is the most major step in this
direction. Rather than having an explicit internal representation, the
parser now parses defn expressions in terms of def and fn.