Correcting bugs and changing representation for alpha-renaming.
Originally drafted alpha-rename to take symbols for the 'old' and
'new' arguments. Realizing this was more pain that it was worth, I
changed this to use the CTCO internal represntation for symbols. This
simplifies the logic involved in alpha-renaming, but puts onus on the
caller to make sure that the inputs are in the correct internal
representation.