Revert "Adding implementations for PGatherFreeVars."
This reverts commit d1fb8efaaf089d7c05f9caaf33a13256f1db0f77.

Traditionally, alpha-renaming requires information about free
variables in order to avoid unintended variable capture. For instance,
alpha-renaming 'y' to 'x' in the expression (lambda (x) y) changes the
meaning of the expression.

However, for the purposes of CTCO, alpha-renaming is intended to be
used internally and renames from an existing symbol to a gensym. By
the very nature of gensym, this means that the free variable analysis
part of alpha-renaming is unnecessary for our purposes.

Due to the fact that the code is mostly complete, save some testing
and kinks to work out, plus the fact that this is the third full
alpha-renamer I've written for this project, the changes reverted here
will be preserved in a separate branch, awaiting the eventuality that
I will think that I yet again need this task done. Hopefully I will
remember that it exists and not reimplement this *yet again.*