Restructuring PLoadTrampoline implementation around record types.
Previously tried to use compile-time metadata to indicate trampoline
entry points in function bodies that had undergone the CPS
transformation. For several reasons, not the least of which being that
Clojure records by nature are immutable and therefore get discarded
often when doing these sorts of transformations, the metadata was
harder to communicate appropriately than initially believed.

To combat this, two new record types were added: TrampMark and
Tramp. The former is introduced during the CPS transformations for
'fn' expressions to indicate which expressions represent points where
the trampoline needs to be loaded, and is replaced by the latter
during the load-tramp pass which inserts the name used for the
trampoline function.

Additionally, the generated code has been updated to include an
initial call to the (generated) trampoline.

Unlike previous versions of CTCO, this ensures that both initially
recursive expressions and calls to functions transformed by the
compiler are appropriately trampolined.