Adding extend-multi macro to replace and improve extend-group.
Previously used the very simple extend-group macro to condense
multiple calls to extend into one. This was well and great, but
required implementations to be expressed as maps.

The extend-multi macro has a much better name and allows for use of
the same syntax as used to implement protocols in defrecord.

Also replacing all instances of extend-group with extend-multi.