Adding named 'fn' support and auto-naming for 'fn' in the parser.
Previously names for 'fn' expressions were added during CPS, as that
was the first point that having a name was necessary. However, the
unrecurify function implemented by PUnRecurify is run before the CPS
transformation and requires an available name for every 'fn'
expression. To this end, support was added to the parser for named
'fn' expressions where only anonymous 'fn' expressions were accepted
before. Additionally, 'fn' expressions without a name are now given an
assigned name in the parser.