Writing first (and incorrect) draft of implementations for PRecurify.
Implementing the recurify function provided by PRecurify for all
relevant CTCO expressions. However, the given implementations are
incorrect. At present they pass the name of the current function and
whether the expression is in tail position, and replaces all
tail-position function applications using that name with the 'recur'
form. However, Clojure introduces the wrinkle in that overloaded
functions may not use recur to call the current function with a
different arity. That is, a given overloaded version of a function may
only recur to itself. Thus the current implementation of PRecurify
will incorrectly insert the 'recur' form for calls to a different
overloaded version of the same function.

For example, consider the following expression:

	(defn foo
	  ([x] (foo x 1))
	  ([x y] (+ x y)))

Given the current implementation for PRecurify, this would be
transformed into this:

	(defn foo
	  ([x] (recur x 1))
	  ([x y] (+ x y)))

Attempting to evaluate this would result in an "incorrect number of
arguments to recur" error. Seeing as though we're in the business of
correctness-preserving transformations and not
incorrectness-introducing ones, this is a bad thing.

The solution to this problem is to pass an additional argument to the
recurify function of PRecurify denoting the arity of the current
function body and only replace function applications when they have
the same name as the current function, are in tail position, _and_
pass the same number of arguments as the current function body
accepts.