Pushing the Recur record into app.clj to avoid cyclic load dependency.
Previously the Recur record type was defined in its own file,
recur.clj. With the advent of the "unrecurify" transformation, which
converts the 'recur' form into a function application, and the
"recurify" tranformation which does the reverse, the App and Recur
record types came to depend on each other in the implementations for
unrecurify and recurify, respectively. To avoid cyclic load
dependencies, the definition for the Recur record type was pushed into
app.clj, the same file in which the definition for the App record type
resides. This seemed the most logical thing to do, as 'recur' is
something of a special form of function application.

It should be noted that this did not resolve the mutual reference
between App and Recur, and so there is the diry, awful (but necessary)
hack of forward declaration of a function which creates the Recur
record type in app.clj.

An alternative solution to this problem is to simple eliminate the
'recur' form as a separate record type and to represent it as an
ordinary function application with a distinguished operator. This
would, however, complicate the implementation of the compiler passes
for the App record type and violate the philosophy of representing the
core language forms via their own record types.