Changing inserted metadata check for 'fn' to use SimpleOpCps.
Previously the inserted test for whether the last argument in the
conflicting arity case of an 'fn' expression used the App record type,
leading to thunkification within the expression. These instances are
being replaced with the SimpleOpCps record type to avoid this. Note
that using this record type is correct given the definition of a
simple operator.